Store build config for homework on linux kernel course.
For ready-to-go build.
We need to have static build on, because our initramfs will not have any shared libraries.
Also we may need to disable tc, because it produces compilation errors on some systems.

The binary will have shared libraries built-in, so if libraries or the busybox itself
were compiled with some -march, that requires more features, than the running CPU has,
you may get illegal instruction errors, while the system boots.
It should work on the same CPU it was compiled on (will work with --enable-kvm in QEMU,
so that QEMU has CPU virtualization access, thus the same CPU features).